rename + add JSDoc + moved component to subfolder
