+++ src/app/core/admin/config-cleanup/configurable-enum-cleanup/configurable-enum-cleanup.component.scss
